requirements.txt
python-dotenv>=1.0
openai>=1.0
tiktoken>=0.5
rich>=13.0
pydantic>=2.0
PyYAML>=6.0
tenacity>=8.0
charset-normalizer>=3.0
pytest>=7.0
